+++ test/scripts/index.js
@@ -3874,6 +3874,9 @@ define(['require'], function (require) { 'use strict';
   function setContextDefaultHref(context, href) {
       return contextManager.setContextDefaultHref(context, href);
   }
+  function getContextDefaultOf(context) {
+      return contextManager.getDefaultOf(context);
+  }
   function getContext(href) {
       if (href === void 0) { href = null; }
       if (href == null) {
@@ -4178,6 +4181,7 @@ define(['require'], function (require) { 'use strict';
       setContext: setContext,
       addContextPath: addContextPath,
       setContextDefaultHref: setContextDefaultHref,
+      getContextDefaultOf: getContextDefaultOf,
       getContext: getContext,
       getHREFs: getHREFs,
       restore: restore,
@@ -4192,6 +4196,7 @@ define(['require'], function (require) { 'use strict';
   var addContextPath_1 = addContextPath;
   var assign_1 = assign;
   var getContext_1 = getContext;
+  var getContextDefaultOf_1 = getContextDefaultOf;
   var getHREFAt_1 = getHREFAt;
   var go_1 = go;
   var index_1 = index;
@@ -4202,12 +4207,13 @@ define(['require'], function (require) { 'use strict';
   var setContextDefaultHref_1 = setContextDefaultHref;
   var start_1 = start;
 
-  var HistoryManager2d438168 = {
+  var HistoryManager377f9dde = {
   	HistoryManager: HistoryManager_1,
   	acquire: acquire_1,
   	addContextPath: addContextPath_1,
   	assign: assign_1,
   	getContext: getContext_1,
+  	getContextDefaultOf: getContextDefaultOf_1,
   	getHREFAt: getHREFAt_1,
   	go: go_1,
   	index: index_1,
@@ -4247,8 +4253,8 @@ define(['require'], function (require) { 'use strict';
       var promise = new Promise(function (resolve) {
           promiseResolve = resolve;
       });
-      HistoryManager2d438168.onWorkFinished(function () {
-          historyLock = HistoryManager2d438168.acquire();
+      HistoryManager377f9dde.onWorkFinished(function () {
+          historyLock = HistoryManager377f9dde.acquire();
           var lock = {
               lock: {
                   get id() {
@@ -4369,7 +4375,7 @@ define(['require'], function (require) { 'use strict';
   var locked_1 = locked;
   var unlock_1 = unlock;
 
-  var NavigationLock98e110ab = {
+  var NavigationLockD1fe8181 = {
   	NavigationLock: NavigationLock_1,
   	lock: lock_1,
   	locked: locked_1,
@@ -4606,7 +4612,7 @@ define(['require'], function (require) { 'use strict';
       if (emit === void 0) { emit = true; }
       var lastEmitRoute = emitRoute;
       emitRoute = emit;
-      return (replace ? HistoryManager2d438168.replace(path) : HistoryManager2d438168.assign(path)).catch(function () {
+      return (replace ? HistoryManager377f9dde.replace(path) : HistoryManager377f9dde.assign(path)).catch(function () {
           emitRoute = lastEmitRoute;
       });
   }
@@ -4704,13 +4710,13 @@ define(['require'], function (require) { 'use strict';
       return main.unroute(path);
   }
   function start(startingContext) {
-      return HistoryManager2d438168.start(startingContext);
+      return HistoryManager377f9dde.start(startingContext);
   }
   function index() {
-      return HistoryManager2d438168.index();
+      return HistoryManager377f9dde.index();
   }
   function getLocationAt(index) {
-      var href = HistoryManager2d438168.getHREFAt(index);
+      var href = HistoryManager377f9dde.getHREFAt(index);
       if (href == null) {
           return null;
       }
@@ -4718,19 +4724,22 @@ define(['require'], function (require) { 'use strict';
   }
   function addContextPath(context, href, isFallback) {
       if (isFallback === void 0) { isFallback = false; }
-      return HistoryManager2d438168.addContextPath(context, href, isFallback);
+      return HistoryManager377f9dde.addContextPath(context, href, isFallback);
   }
   function setContextDefaultHref(context, href) {
-      return HistoryManager2d438168.setContextDefaultHref(context, href);
+      return HistoryManager377f9dde.setContextDefaultHref(context, href);
   }
   function setContext(context) {
-      return HistoryManager2d438168.setContext(context);
+      return HistoryManager377f9dde.setContext(context);
   }
   function getContext(href) {
-      return HistoryManager2d438168.getContext(href);
+      return HistoryManager377f9dde.getContext(href);
   }
   function restoreContext(context, defaultHref) {
-      return HistoryManager2d438168.restore(context);
+      return HistoryManager377f9dde.restore(context);
+  }
+  function getContextDefaultOf(context) {
+      return HistoryManager377f9dde.getContextDefaultOf(context);
   }
   function emit(single) {
       if (single === void 0) { single = false; }
@@ -4764,7 +4773,7 @@ define(['require'], function (require) { 'use strict';
           else {
               var lastEmitRoute_1 = emitRoute;
               emitRoute = options.emit == null ? true : options.emit;
-              HistoryManager2d438168.go(path_index).then(promiseResolve, function () {
+              HistoryManager377f9dde.go(path_index).then(promiseResolve, function () {
                   emitRoute = lastEmitRoute_1;
               });
           }
@@ -4773,7 +4782,7 @@ define(['require'], function (require) { 'use strict';
   function setQueryParam(param, value, options) {
       var promiseResolve;
       var promise = new Promise(function (resolve) { promiseResolve = resolve; });
-      HistoryManager2d438168.onWorkFinished(function () {
+      HistoryManager377f9dde.onWorkFinished(function () {
           var location = getLocation();
           if (value === undefined) {
               location.removeQueryParam(param);
@@ -4786,11 +4795,11 @@ define(['require'], function (require) { 'use strict';
       return promise;
   }
   function lock() {
-      return NavigationLock98e110ab.lock();
+      return NavigationLockD1fe8181.lock();
   }
   function unlock(force) {
       if (force === void 0) { force = true; }
-      return NavigationLock98e110ab.unlock(force);
+      return NavigationLockD1fe8181.unlock(force);
   }
   function destroy() {
       throw new Error("cannot destroy main Router");
@@ -4803,7 +4812,7 @@ define(['require'], function (require) { 'use strict';
       _emit();
   }
   function isLocked() {
-      return NavigationLock98e110ab.locked();
+      return NavigationLockD1fe8181.locked();
   }
 
   var Router = /*#__PURE__*/Object.freeze({
@@ -4821,6 +4830,7 @@ define(['require'], function (require) { 'use strict';
       setContext: setContext,
       getContext: getContext,
       restoreContext: restoreContext,
+      getContextDefaultOf: getContextDefaultOf,
       emit: emit,
       create: create,
       go: go,
@@ -4831,7 +4841,7 @@ define(['require'], function (require) { 'use strict';
       getBase: getBase,
       setBase: setBase,
       isLocked: isLocked,
-      NavigationLock: NavigationLock98e110ab.NavigationLock
+      NavigationLock: NavigationLockD1fe8181.NavigationLock
   });
 
   var locks = [];
@@ -4952,8 +4962,8 @@ define(['require'], function (require) { 'use strict';
   exports.OptionsManager = OptionsManager3fd4d9f6.OptionsManager;
   exports.ContextManager = ContextManager6ca49066.ContextManager$1;
   exports.URLManager = ContextManager6ca49066.URLManager;
-  exports.HistoryManager = HistoryManager2d438168.HistoryManager;
-  exports.NavigationLock = NavigationLock98e110ab.NavigationLock;
+  exports.HistoryManager = HistoryManager377f9dde.HistoryManager;
+  exports.NavigationLock = NavigationLockD1fe8181.NavigationLock;
   exports.Router = Router;
   exports.WorkManager = WorkManager;
   });
@@ -5023,6 +5033,7 @@ define(['require'], function (require) { 'use strict';
           cancelAnimationFrame(nextFrame);
       }
       var lastTime;
+      var eventDispatched = false;
       var step = function () {
           if (loadingDone && loadingProgress === 5) {
               loadingProgress = 100;
@@ -5033,10 +5044,13 @@ define(['require'], function (require) { 'use strict';
           var last = lastTime;
           var delta = ((lastTime = Date.now()) - last);
           if (loadingProgress >= 100) {
+              if (!eventDispatched) {
+                  window.dispatchEvent(new Event("routerload"));
+                  eventDispatched = true;
+              }
               if ((doneTime -= delta) <= 0) {
                   doneTime = visibilityTime;
                   loadingBarContainer.style.display = "none";
-                  window.dispatchEvent(new Event("routerload"));
               }
               else {
                   requestAnimationFrame(step);
@@ -5414,7 +5428,9 @@ define(['require'], function (require) { 'use strict';
 
       href(toA = true) {
           if (typeof this.props.href !== "string") {
-              return null;
+              const context = this.context();
+              console.log(context, cjs.Router.getContextDefaultOf(context));
+              return context != null ? cjs.Router.getContextDefaultOf(context) : null;
           }
           if (this._href == null) {
               this._href = cjs.Router.getLocation().hrefIf(this.props.href);
@@ -5443,7 +5459,7 @@ define(['require'], function (require) { 'use strict';
             'name': 'href',
 
             'evaluate': function(scope) {
-              return scope.href();
+              return "#" + scope.href();
             }
           }, {
             'type': expressionTypes.ATTRIBUTE,
@@ -5566,12 +5582,6 @@ define(['require'], function (require) { 'use strict';
           cjs.Router.start("home").then(() => console.log("started"));
       },
 
-      restore(context) {
-          return () => {
-              cjs.Router.restoreContext(context);
-          };
-      },
-
       components: {
           homepage: lazy(() => new Promise(function (resolve, reject) { require(['./homepage-eebb63b5'], resolve, reject) }))
       }
@@ -5579,7 +5589,7 @@ define(['require'], function (require) { 'use strict';
 
     'template': function(template, expressionTypes, bindingTypes, getComponent) {
       return template(
-        '<navigate expr21="expr21" href="home"></navigate><navigate expr22="expr22" href="me"></navigate><navigate expr23="expr23"></navigate><div></div><button expr24="expr24">Restore home</button><button expr25="expr25">Restore profile</button><div></div><router expr26="expr26"></router>',
+        '<navigate expr21="expr21" href="home"></navigate><navigate expr22="expr22" href="me"></navigate><navigate expr23="expr23"></navigate><div></div><navigate expr24="expr24" context="home"></navigate><navigate expr25="expr25" context="profile"></navigate><div></div><router expr26="expr26"></router>',
         [{
           'type': bindingTypes.TAG,
           'getComponent': getComponent,
@@ -5640,29 +5650,39 @@ define(['require'], function (require) { 'use strict';
           'redundantAttribute': 'expr23',
           'selector': '[expr23]'
         }, {
+          'type': bindingTypes.TAG,
+          'getComponent': getComponent,
+
+          'evaluate': function(scope) {
+            return 'navigate';
+          },
+
+          'slots': [{
+            'id': 'default',
+            'html': 'Restore home',
+            'bindings': []
+          }],
+
+          'attributes': [],
           'redundantAttribute': 'expr24',
-          'selector': '[expr24]',
-
-          'expressions': [{
-            'type': expressionTypes.EVENT,
-            'name': 'onclick',
-
-            'evaluate': function(scope) {
-              return scope.restore("home");
-            }
-          }]
+          'selector': '[expr24]'
         }, {
+          'type': bindingTypes.TAG,
+          'getComponent': getComponent,
+
+          'evaluate': function(scope) {
+            return 'navigate';
+          },
+
+          'slots': [{
+            'id': 'default',
+            'html': 'Restore profile',
+            'bindings': []
+          }],
+
+          'attributes': [],
           'redundantAttribute': 'expr25',
-          'selector': '[expr25]',
-
-          'expressions': [{
-            'type': expressionTypes.EVENT,
-            'name': 'onclick',
-
-            'evaluate': function(scope) {
-              return scope.restore("profile");
-            }
-          }]
+          'selector': '[expr25]'
         }, {
           'type': bindingTypes.TAG,
           'getComponent': getComponent,